Add isValid field to InputPktsData (pass-through)
